requirements.txt
slack-sdk>=3.27.0
fastapi>=0.115.0
uvicorn>=0.30.0
apscheduler>=3.10.0
jinja2>=3.1.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
langfuse>=3.0.0
